Make CPKColorWell and CPKPopover tolerate the color well being removed from the view hierarchy while the popover is open. This is necessary to show a color well in a table view.

+++ ColorPicker/ColorPickerTests/ColorPickerTests.m
@@ -36,7 +36,7 @@ const NSInteger kMaxValue = 10;
 }
 
 - (NSArray *)randomPointsWithSeed:(NSInteger)seed {
-    srand(seed);
+    srand((unsigned int)seed);
     const NSInteger kMaxPoints = 50;
     NSInteger numPoints = 1 + abs(rand()) % kMaxPoints;
     NSMutableArray *points = [NSMutableArray array];
